backend/requirements.txt
# flask
django
gunicorn
pathlib
pyyaml
markdown